// sdk/typescript/src/models/ListPassportsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Pagination } from './Pagination';
import type { Passport } from './Passport';
export type ListPassportsResponse = {
    success: boolean;
    passports: Array<Passport>;
    pagination: Pagination;
};